move form logic out of page and into testable components
