+++ components/MashForms/ContactForm.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ContactForm = (): React.ReactElement => {
+  return <h1>Work on contact</h1>;
+};
+
+export default ContactForm;
